Refactored (java-application): Splitted GraphQL-Module into interface and implementation
