+++ server/src/main/java/GraphQL/Fetcher/MainGraphQLDataFetchers.java
@@ -1,0 +1,8 @@
+package GraphQL.Fetcher;
+
+import graphql.schema.DataFetcher;
+
+public interface MainGraphQLDataFetchers
+{
+	DataFetcher searchForFileMetadataFetcher();
+}
