+++ server/src/main/java/GraphQL/GraphQLModule.java
@@ -1,6 +1,10 @@
 package GraphQL;
 
 import Database.Database;
+import GraphQL.Fetcher.Impl.MainGraphQLDataFetchersImpl;
+import GraphQL.Fetcher.MainGraphQLDataFetchers;
+import GraphQL.Provider.Impl.MainGraphQLProviderImpl;
+import GraphQL.Provider.MainGraphQLProvider;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
@@ -10,26 +14,19 @@ import java.io.IOException;
 
 public class GraphQLModule extends AbstractModule
 {
-//    @Provides
-//    @Singleton
-//    static GraphQLDataFetchers provideGraphQLDataFetchers(DatabaseProvider databaseProvider)
-//    {
-//        return new GraphQLDataFetchers(databaseProvider);
-//    }
 
     @Override
     protected void configure()
     {
-       // this.bind(GraphQLDataFetchers.class).to(GraphQLDataFetchers.class).in(Singleton.class);
+        this.bind(MainGraphQLDataFetchers.class).to(MainGraphQLDataFetchersImpl.class).in(Singleton.class);
     }
-
 
 
     @Provides
     @Singleton
-    static GraphQL provideGraphQL(MainGraphQLDataFetchers mainGraphQLDataFetchers, Database database) throws IOException
+    static GraphQL provideGraphQL(MainGraphQLDataFetchers mainGraphQLDataFetchers) throws IOException
     {
-        return new MainGraphQLProvider(mainGraphQLDataFetchers, database).init().getGraphQL();
+        return new MainGraphQLProviderImpl(mainGraphQLDataFetchers).init().getGraphQL();
     }
